+++ webpack.config.ts
@@ -1,0 +1,20 @@
+import type { Configuration } from 'webpack';
+import { merge } from 'webpack-merge';
+import grafanaConfig from './.config/webpack/webpack.config';
+
+const config = async (env): Promise<Configuration> => {
+  const baseConfig = await grafanaConfig(env);
+
+  return merge(baseConfig, {
+    // update output configuration
+    // other configurations stay the same
+    output: {
+      ...baseConfig.output,
+      clean: {
+        keep: new RegExp(`(.*?_(amd64|arm(64)?|s390x)(.exe)?|go_plugin_build_manifest)`),
+      },
+    },
+  });
+};
+
+export default config;