+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## tip
 
-## v0.16.0
-
 * FEATURE: implements the getLogRowContext method. See [this PR](https://github.com/VictoriaMetrics/victorialogs-datasource/pull/247).
+* FEATURE: add support for build the linux/s390x. Extend backend build process to add more architectures. See [this issue](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/250).
+* FEATURE: add configuration screen for Custom query parameters. See [this issue](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/252).
+
+* BUGFIX: fix build annotation from the label field. All labels transforms to the string representation. See [this issue](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/188).
+
 
 ## v0.15.0